TECHDEBT: Created UI Folder and reorganized reutilizable components into it
