+++ src/components/UI/Button/Button.module.scss
@@ -1,4 +1,4 @@
-@import "../../assets/css/colors.scss";
+@import "../../../assets/css/colors.scss";
 
 .button {
   background-color: $background-color-transparent;
